Refactor capture sync architecture
