+++ Symi/Sources/Core/Episodes/EpisodeCore.swift
@@ -1,5 +1,4 @@
 import Foundation
-import Observation
 
 struct WeatherRecord: Equatable, Sendable {
     nonisolated let recordedAt: Date
@@ -223,13 +222,6 @@ struct CustomMedicationDefinitionDraft: Equatable, Sendable {
     nonisolated let dosage: String
 }
 
-struct EpisodeEditorMedicationGroup: Identifiable, Equatable, Sendable {
-    nonisolated let id: String
-    nonisolated let title: String
-    nonisolated let footer: String?
-    nonisolated let items: [MedicationDefinitionRecord]
-}
-
 protocol EpisodeRepository: Sendable {
     nonisolated func fetchRecent() throws -> [EpisodeRecord]
     nonisolated func fetchByDay(_ day: Date) throws -> [EpisodeRecord]
@@ -308,497 +300,5 @@ enum EpisodeSaveError: LocalizedError {
         case .futureDate:
             "Eine Episode kann nicht in der Zukunft erfasst werden."
         }
-    }
-}
-
-enum WeatherLoadState: Equatable {
-    case idle
-    case loading
-    case loaded(WeatherSnapshotData)
-    case unavailable(String)
-}
-
-@MainActor
-@Observable
-final class EpisodeEditorController {
-    let mode: EpisodeEditorMode
-    let symptomOptions = [
-        "Übelkeit",
-        "Lichtempfindlichkeit",
-        "Geräuschempfindlichkeit",
-        "Aura",
-        "Kiefer-/Aufbissschmerz",
-        "Pochen, Pulsieren"
-    ]
-    let triggerOptions = ["Stress", "Schlafmangel", "Alkohol", "Menstruation", "Bildschirmzeit"]
-
-    var draft: EpisodeDraft
-    var medicationSearchText = ""
-    var saveMessageVisible = false
-    var isSaving = false
-    var validationMessage: String?
-    var weatherLoadState: WeatherLoadState = .idle
-    var customMedicationEditor: CustomMedicationEditorSheetState?
-    var pendingMedicationDeletion: MedicationDefinitionRecord?
-    private(set) var medicationDefinitions: [MedicationDefinitionRecord] = []
-    private var selectedMedicationKeys: Set<String> = []
-    private var medicationSelectionIndicesByKey: [String: Int] = [:]
-    private var selectedMedicationsCache: [MedicationSelectionDraft] = []
-    private var allMedicationGroupsCache: [EpisodeEditorMedicationGroup] = []
-
-    private let saveEpisodeUseCase: SaveEpisodeUseCase
-    private let episodeRepository: EpisodeRepository
-    private let medicationRepository: MedicationCatalogRepository
-    private let weatherService: WeatherService
-    private let locationService: LocationService
-    private let healthService: HealthService
-    private var originalStartedAt: Date?
-    private var originalWeatherSnapshot: WeatherSnapshotData?
-
-    init(
-        episodeID: UUID?,
-        initialStartedAt: Date?,
-        episodeRepository: EpisodeRepository,
-        medicationRepository: MedicationCatalogRepository,
-        weatherService: WeatherService,
-        locationService: LocationService,
-        healthService: HealthService
-    ) {
-        self.episodeRepository = episodeRepository
-        self.medicationRepository = medicationRepository
-        self.weatherService = weatherService
-        self.locationService = locationService
-        self.healthService = healthService
-        self.saveEpisodeUseCase = SaveEpisodeUseCase(repository: episodeRepository)
-
-        if let episodeID {
-            self.mode = .edit
-            self.draft = EpisodeDraft.makeNew(initialStartedAt: initialStartedAt)
-            self.originalStartedAt = nil
-            self.originalWeatherSnapshot = nil
-            Task { await loadEpisode(id: episodeID) }
-        } else {
-            self.mode = .create
-            if AppStoreScreenshotMode.isEnabled {
-                let screenshotDraft = AppStoreScreenshotMode.sampleDraft(initialStartedAt: initialStartedAt)
-                self.draft = screenshotDraft
-                self.weatherLoadState = .loaded(AppStoreScreenshotMode.sampleWeatherSnapshot(for: screenshotDraft.startedAt))
-            } else {
-                self.draft = EpisodeDraft.makeNew(initialStartedAt: initialStartedAt)
-            }
-            self.originalStartedAt = nil
-            self.originalWeatherSnapshot = nil
-        }
-
-        rebuildMedicationCaches()
-        reloadMedicationDefinitions()
-    }
-
-    var filteredMedicationGroups: [EpisodeEditorMedicationGroup] {
-        let query = medicationSearchText.trimmingCharacters(in: .whitespacesAndNewlines)
-        let groups = allMedicationGroupsCache
-
-        guard !query.isEmpty else {
-            return groups
-        }
-
-        return groups.compactMap { group in
-            let items = group.items.filter { $0.name.localizedCaseInsensitiveContains(query) }
-            guard !items.isEmpty else {
-                return nil
-            }
-
-            return EpisodeEditorMedicationGroup(
-                id: group.id,
-                title: group.title,
-                footer: group.footer,
-                items: items
-            )
-        }
-    }
-
-    var selectedMedications: [MedicationSelectionDraft] {
-        selectedMedicationsCache
-    }
-
-    func reloadMedicationDefinitions() {
-        let repository = medicationRepository
-        Task {
-            let definitions = await PerformanceInstrumentation.measure("MedicationDefinitionsReload") {
-                await Task.detached(priority: .userInitiated) {
-                    PerformanceInstrumentation.measure("MedicationRepositoryFetchDefinitions") {
-                        (try? repository.fetchDefinitions(searchText: nil)) ?? []
-                    }
-                }.value
-            }
-            medicationDefinitions = definitions
-            rebuildMedicationCaches()
-        }
-    }
-
-    private func loadEpisode(id: UUID) async {
-        let record = await PerformanceInstrumentation.measure("EpisodeEditorLoad") {
-            let repository = episodeRepository
-            return await Task.detached(priority: .userInitiated) {
-                PerformanceInstrumentation.measure("EpisodeRepositoryLoad") {
-                    try? repository.load(id: id)
-                }
-            }.value
-        }
-
-        guard let record else {
-            return
-        }
-
-        draft = EpisodeDraft.from(record: record)
-        originalStartedAt = record.startedAt
-        originalWeatherSnapshot = record.weather.map(WeatherSnapshotData.init)
-        weatherLoadState = record.weather.map { .loaded(WeatherSnapshotData(record: $0)) } ?? .idle
-        rebuildMedicationCaches()
-    }
-
-    func save(onSaved: (() -> Void)?, onDismiss: @escaping () -> Void) {
-        guard !isSaving else {
-            return
-        }
-
-        validationMessage = nil
-        isSaving = true
-
-        Task {
-            await PerformanceInstrumentation.measure("EpisodeEditorSave") {
-                defer { isSaving = false }
-
-                do {
-                    let weatherSnapshot = try await weatherSnapshotForSave()
-                    let healthContext = await healthContextForSave()
-                    let savedID = try await saveEpisodeUseCase.execute(draft, weatherSnapshot: weatherSnapshot, healthContext: healthContext)
-                    await writeHealthSampleIfNeeded(episodeID: savedID)
-                    reloadMedicationDefinitions()
-                    validationMessage = nil
-
-                    if mode == .create, onSaved == nil {
-                        draft = EpisodeDraft.makeNew()
-                        medicationSearchText = ""
-                        customMedicationEditor = nil
-                        pendingMedicationDeletion = nil
-                        weatherLoadState = .idle
-                        rebuildMedicationCaches()
-                        saveMessageVisible = true
-                    } else {
-                        onSaved?()
-                        onDismiss()
-                    }
-                } catch {
-                    validationMessage = error.localizedDescription
-                }
-            }
-        }
-    }
-
-    func refreshWeather() async {
-        await PerformanceInstrumentation.measure("EpisodeWeatherRefresh") {
-            if AppStoreScreenshotMode.isEnabled {
-                weatherLoadState = .loaded(AppStoreScreenshotMode.sampleWeatherSnapshot(for: draft.startedAt))
-                return
-            }
-
-            if draft.startedAt > .now {
-                weatherLoadState = .unavailable("Für zukünftige Zeitpunkte wird kein Wetter geladen.")
-                return
-            }
-
-            if isStartedAtUnchanged, let originalWeatherSnapshot {
-                weatherLoadState = .loaded(originalWeatherSnapshot)
-                return
-            }
-
-            weatherLoadState = .loading
-
-            do {
-                let location = try await locationService.requestApproximateLocation()
-                guard let snapshot = try await weatherService.fetchWeather(for: draft.startedAt, location: location) else {
-                    weatherLoadState = .unavailable("Für diesen Zeitpunkt konnten keine Wetterdaten geladen werden.")
-                    return
-                }
-                weatherLoadState = .loaded(snapshot)
-            } catch let error as EpisodeSaveError {
-                weatherLoadState = .unavailable(error.localizedDescription)
-            } catch let error as any LocalizedError {
-                weatherLoadState = .unavailable(error.errorDescription ?? "Wetterdaten konnten nicht geladen werden.")
-            } catch {
-                weatherLoadState = .unavailable("Wetterdaten konnten nicht geladen werden.")
-            }
-        }
-    }
-
-    func isMedicationSelected(_ definition: MedicationDefinitionRecord) -> Bool {
-        selectedMedicationKeys.contains(definition.selectionKey)
-    }
-
-    func quantity(for definition: MedicationDefinitionRecord) -> Int {
-        guard let index = medicationSelectionIndicesByKey[definition.selectionKey] else {
-            return 1
-        }
-
-        return draft.medications[index].quantity
-    }
-
-    func toggleMedicationSelection(for definition: MedicationDefinitionRecord) {
-        if let index = medicationSelectionIndicesByKey[definition.selectionKey] {
-            draft.medications[index].isSelected.toggle()
-            draft.medications[index].quantity = max(1, draft.medications[index].quantity)
-        } else {
-            draft.medications.append(MedicationSelectionDraft(definition: definition))
-        }
-        rebuildMedicationCaches()
-    }
-
-    func incrementMedicationQuantity(for definition: MedicationDefinitionRecord) {
-        if let index = medicationSelectionIndicesByKey[definition.selectionKey] {
-            draft.medications[index].quantity += 1
-            draft.medications[index].isSelected = true
-        } else {
-            draft.medications.append(MedicationSelectionDraft(definition: definition))
-        }
-        rebuildMedicationCaches()
-    }
-
-    func decrementMedicationQuantity(for definition: MedicationDefinitionRecord) {
-        guard let index = medicationSelectionIndicesByKey[definition.selectionKey] else {
-            return
-        }
-
-        draft.medications[index].quantity = max(1, draft.medications[index].quantity - 1)
-        draft.medications[index].isSelected = true
-        rebuildMedicationCaches()
-    }
-
-    func removeMedicationSelection(id: UUID) {
-        guard let index = draft.medications.firstIndex(where: { $0.id == id }) else {
-            return
-        }
-
-        draft.medications[index].isSelected = false
-        draft.medications[index].quantity = 1
-        rebuildMedicationCaches()
-    }
-
-    func presentEditor(for definition: MedicationDefinitionRecord?) {
-        customMedicationEditor = CustomMedicationEditorSheetState(definition: definition)
-    }
-
-    func saveCustomMedication(from draft: CustomMedicationDefinitionDraft) {
-        let trimmedName = draft.name.trimmingCharacters(in: .whitespacesAndNewlines)
-
-        guard !trimmedName.isEmpty else {
-            validationMessage = "Bitte gib einen Namen für das eigene Medikament ein."
-            return
-        }
-
-        if let existing = medicationDefinitions.first(where: {
-            $0.catalogKey != draft.id &&
-            $0.name.compare(trimmedName, options: [.caseInsensitive, .diacriticInsensitive]) == .orderedSame
-        }) {
-            toggleMedicationSelection(for: existing)
-            customMedicationEditor = nil
-            validationMessage = nil
-            return
-        }
-
-        let repository = medicationRepository
-        Task {
-            do {
-                let definition = try await Task.detached(priority: .userInitiated) {
-                    try repository.saveCustomDefinition(draft)
-                }.value
-                reloadMedicationDefinitions()
-                customMedicationEditor = nil
-                validationMessage = nil
-
-                if let existingSelectionKey = draft.originalSelectionKey {
-                    updateMedicationSelection(from: existingSelectionKey, to: definition)
-                } else {
-                    toggleMedicationSelection(for: definition)
-                }
-            } catch {
-                validationMessage = "Eigenes Medikament konnte nicht gespeichert werden."
-            }
-        }
-    }
-
-    func deleteCustomMedication(_ definition: MedicationDefinitionRecord) {
-        draft.medications.removeAll { $0.selectionKey == definition.selectionKey }
-        rebuildMedicationCaches()
-        let repository = medicationRepository
-        Task {
-            do {
-                try await Task.detached(priority: .userInitiated) {
-                    try repository.softDeleteCustomDefinition(catalogKey: definition.catalogKey)
-                }.value
-                reloadMedicationDefinitions()
-                pendingMedicationDeletion = nil
-                validationMessage = nil
-            } catch {
-                validationMessage = "Eigenes Medikament konnte nicht gelöscht werden."
-            }
-        }
-    }
-
-    private func rebuildMedicationCaches() {
-        rebuildMedicationSelectionCache()
-        rebuildMedicationGroupCache()
-    }
-
-    private func rebuildMedicationSelectionCache() {
-        medicationSelectionIndicesByKey = draft.medications.enumerated().reduce(into: [:]) { result, item in
-            let (index, medication) = item
-            result[medication.selectionKey] = index
-        }
-        selectedMedicationKeys = Set(draft.medications.filter(\.isSelected).map(\.selectionKey))
-        selectedMedicationsCache = draft.medications
-            .filter(\.isSelected)
-            .sorted { $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending }
-    }
-
-    private func rebuildMedicationGroupCache() {
-        let knownKeys = Set(medicationDefinitions.map(\.selectionKey))
-        let persistedGroups = Dictionary(grouping: medicationDefinitions) { $0.groupID }
-        let sortedGroupIDs = persistedGroups.keys.sorted { lhs, rhs in
-            let leftOrder = persistedGroups[lhs]?.map(\.sortOrder).min() ?? .max
-            let rightOrder = persistedGroups[rhs]?.map(\.sortOrder).min() ?? .max
-            return leftOrder < rightOrder
-        }
-
-        var groups = sortedGroupIDs.compactMap { groupID -> EpisodeEditorMedicationGroup? in
-            guard let items = persistedGroups[groupID], let first = items.first else {
-                return nil
-            }
-
-            return EpisodeEditorMedicationGroup(
-                id: groupID,
-                title: first.groupTitle,
-                footer: first.groupFooter,
-                items: items.sorted { $0.sortOrder < $1.sortOrder }
-            )
-        }
-
-        let orphanSelections = draft.medications.compactMap { selection -> MedicationDefinitionRecord? in
-            guard !knownKeys.contains(selection.selectionKey) else {
-                return nil
-            }
-
-            return MedicationDefinitionRecord(
-                catalogKey: "selection:\(selection.selectionKey)",
-                groupID: "custom-medications",
-                groupTitle: "Eigene Medikamente",
-                groupFooter: "Eigene Medikamente werden lokal in SwiftData gespeichert und bleiben in deiner persönlichen Auswahlliste verfügbar.",
-                name: selection.name,
-                category: selection.category,
-                suggestedDosage: selection.dosage,
-                sortOrder: Int.max - 1,
-                isCustom: true,
-                isDeleted: false
-            )
-        }
-
-        if !orphanSelections.isEmpty {
-            groups.removeAll { $0.id == "custom-medications" }
-            let customItems = (persistedGroups["custom-medications"] ?? []) + orphanSelections
-            let deduped = Dictionary(uniqueKeysWithValues: customItems.map { ($0.selectionKey, $0) }).values
-                .sorted { $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending }
-
-            groups.append(
-                EpisodeEditorMedicationGroup(
-                    id: "custom-medications",
-                    title: "Eigene Medikamente",
-                    footer: "Eigene Medikamente werden lokal in SwiftData gespeichert und bleiben in deiner persönlichen Auswahlliste verfügbar.",
-                    items: deduped
-                )
-            )
-        }
-
-        allMedicationGroupsCache = groups
-    }
-
-    private func updateMedicationSelection(from oldSelectionKey: String, to definition: MedicationDefinitionRecord) {
-        guard let index = draft.medications.firstIndex(where: { $0.selectionKey == oldSelectionKey }) else {
-            return
-        }
-
-        draft.medications[index].selectionKey = definition.selectionKey
-        draft.medications[index].name = definition.name
-        draft.medications[index].category = definition.category
-        draft.medications[index].dosage = definition.suggestedDosage
-        draft.medications[index].isSelected = true
-        rebuildMedicationCaches()
-    }
-
-    private var isStartedAtUnchanged: Bool {
-        guard let originalStartedAt else {
-            return false
-        }
-
-        return originalStartedAt == draft.startedAt
-    }
-
-    private func weatherSnapshotForSave() async throws -> WeatherSnapshotData? {
-        if draft.startedAt > .now {
-            throw EpisodeSaveError.futureDate
-        }
-
-        if isStartedAtUnchanged {
-            return originalWeatherSnapshot
-        }
-
-        switch weatherLoadState {
-        case .loaded(let snapshot):
-            return snapshot
-        case .idle, .loading, .unavailable:
-            await refreshWeather()
-            if case .loaded(let snapshot) = weatherLoadState {
-                return snapshot
-            }
-            return nil
-        }
-    }
-
-    private func healthContextForSave() async -> HealthContextSnapshotData? {
-        do {
-            return try await healthService.contextSnapshot(for: draft)
-        } catch {
-            return nil
-        }
-    }
-
-    private func writeHealthSampleIfNeeded(episodeID: UUID) async {
-        do {
-            try await healthService.writeEpisode(id: episodeID, draft: draft)
-        } catch {}
-    }
-}
-
-enum EpisodeEditorMode {
-    case create
-    case edit
-}
-
-struct CustomMedicationEditorSheetState: Identifiable, Equatable {
-    let id: String
-    let originalSelectionKey: String?
-    let initialName: String
-    let initialCategory: MedicationCategory
-    let initialDosage: String
-
-    init(definition: MedicationDefinitionRecord?) {
-        id = definition?.catalogKey ?? UUID().uuidString
-        originalSelectionKey = definition?.selectionKey
-        initialName = definition?.name ?? ""
-        initialCategory = definition?.category ?? .other
-        initialDosage = definition?.suggestedDosage ?? ""
-    }
-
-    var isEditing: Bool {
-        originalSelectionKey != nil
     }
 }
